+++ src/plus/integrations/integration.ts
@@ -16,6 +16,7 @@ import type {
 	PullRequestState,
 	SearchedPullRequest,
 } from '../../git/models/pullRequest';
+import type { PullRequestURLIdentity } from '../../git/models/pullRequest.utils';
 import type { RepositoryMetadata } from '../../git/models/repositoryMetadata';
 import { showIntegrationDisconnectedTooManyFailedRequestsWarningMessage } from '../../messages';
 import { gate } from '../../system/decorators/gate';
@@ -1292,4 +1293,6 @@ export abstract class HostingIntegration<
 		repos?: T[],
 		cancellation?: CancellationToken,
 	): Promise<PullRequest[] | undefined>;
+
+	getPullRequestIdentityValuesFromSearch?(search: string): PullRequestURLIdentity | undefined;
 }