+++ src/plus/launchpad/launchpad.ts
@@ -43,7 +43,7 @@ import type { LaunchpadTelemetryContext, Source, Sources, TelemetryEvents } from
 import type { Container } from '../../container';
 import { PlusFeatures } from '../../features';
 import { doesPullRequestSatisfyRepositoryURLIdentity } from '../../git/models/pullRequest';
-import { getPullRequestIdentityValuesFromSearch } from '../../git/models/pullRequest.utils';
+import type { PullRequestURLIdentity } from '../../git/models/pullRequest.utils';
 import type { QuickPickItemOfT } from '../../quickpicks/items/common';
 import { createQuickPickItemOfT, createQuickPickSeparator } from '../../quickpicks/items/common';
 import type { DirectiveQuickPickItem } from '../../quickpicks/items/directive';
@@ -599,14 +599,10 @@ export class LaunchpadCommand extends QuickCommand<State> {
 					return true;
 				}
 
-				// TODO: This needs to be generalized to work outside of GitHub,
-				// The current idea is that we should iterate the connected integrations and apply their parsing.
-				// Probably we even want to build a map like this: { integrationId: identity }
-				// Then when we iterate local items we can check them to corresponding identitie according to the item's repo type.
-				// Same with API: we iterate connected integrations and search in each of them with the corresponding identity.
-				const prUrlIdentity = getPullRequestIdentityValuesFromSearch(value);
+				const prUrlIdentity: { [key in string]?: PullRequestURLIdentity } =
+					await this.container.launchpad.getPullRequestIdentityValuesFromSearch(value);
 
-				if (prUrlIdentity.prNumber != null) {
+				if (Object.keys(prUrlIdentity).length > 0) {
 					// We can identify the PR number, so let's try to find it locally:
 					const launchpadItems = quickpick.items.filter((i): i is LaunchpadItemQuickPickItem => 'item' in i);
 					let item = launchpadItems.find(i =>
@@ -615,7 +611,7 @@ export class LaunchpadCommand extends QuickCommand<State> {
 					);
 					if (item == null) {
 						// Haven't found full match, so let's at least find something with the same pr number
-						item = launchpadItems.find(i => i.item.id === prUrlIdentity.prNumber);
+						item = launchpadItems.find(i => i.item.id === prUrlIdentity[i.item.provider.id]?.prNumber);
 					}
 					if (item != null) {
 						if (!item.alwaysShow) {
@@ -767,6 +763,11 @@ export class LaunchpadCommand extends QuickCommand<State> {
 				createQuickPickSeparator('Actions'),
 			];
 
+			const checkoutable =
+				state.item.type === 'pullrequest' &&
+				state.item.headRef != null &&
+				state.item.repoIdentity?.remote?.url != null;
+
 			for (const action of state.item.suggestedActions) {
 				switch (action) {
 					case 'merge': {
@@ -814,41 +815,50 @@ export class LaunchpadCommand extends QuickCommand<State> {
 							),
 						);
 						break;
-					case 'switch':
-						confirmations.push(
-							createQuickPickItemOfT(
-								{
-									label: 'Switch to Branch',
-									detail: 'Will checkout the branch, create or open a worktree',
-								},
-								action,
-							),
-						);
+					case 'switch': {
+						if (checkoutable) {
+							confirmations.push(
+								createQuickPickItemOfT(
+									{
+										label: 'Switch to Branch',
+										detail: 'Will checkout the branch, create or open a worktree',
+									},
+									action,
+								),
+							);
+						}
 						break;
-					case 'open-worktree':
-						confirmations.push(
-							createQuickPickItemOfT(
-								{
-									label: 'Open in Worktree',
-									detail: 'Will create or open a worktree in a new window',
-								},
-								action,
-							),
-						);
+					}
+					case 'open-worktree': {
+						if (checkoutable) {
+							confirmations.push(
+								createQuickPickItemOfT(
+									{
+										label: 'Open in Worktree',
+										detail: 'Will create or open a worktree in a new window',
+									},
+									action,
+								),
+							);
+						}
 						break;
-					case 'switch-and-code-suggest':
-						confirmations.push(
-							createQuickPickItemOfT(
-								{
-									label: `Switch & Suggest ${
-										state.item.viewer.isAuthor ? 'Additional ' : ''
-									}Code Changes`,
-									detail: 'Will checkout and start suggesting code changes',
-								},
-								action,
-							),
-						);
+					}
+					case 'switch-and-code-suggest': {
+						if (checkoutable) {
+							confirmations.push(
+								createQuickPickItemOfT(
+									{
+										label: `Switch & Suggest ${
+											state.item.viewer.isAuthor ? 'Additional ' : ''
+										}Code Changes`,
+										detail: 'Will checkout and start suggesting code changes',
+									},
+									action,
+								),
+							);
+						}
 						break;
+					}
 					case 'code-suggest':
 						confirmations.push(
 							createQuickPickItemOfT(
@@ -882,16 +892,19 @@ export class LaunchpadCommand extends QuickCommand<State> {
 							),
 						);
 						break;
-					case 'open-in-graph':
-						confirmations.push(
-							createQuickPickItemOfT(
-								{
-									label: 'Open in Commit Graph',
-								},
-								action,
-							),
-						);
+					case 'open-in-graph': {
+						if (checkoutable) {
+							confirmations.push(
+								createQuickPickItemOfT(
+									{
+										label: 'Open in Commit Graph',
+									},
+									action,
+								),
+							);
+						}
 						break;
+					}
 				}
 			}